Close #4 - Use vanilla traits to modify the outcome of lockpicking
The "Lucky" trait will increase the chance of opening a lock
successfully, whereas "Unlucky" will decrease it. The "Clumsy" and
"Graceful" traits affect the noise the player creates while fiddling
with a lock.

"KeenHearing" and "HardOfHearing" only affect the success of the
lockpicking action.

The chance for breaking a bobby pin in the lockpicking process is now
also affected by these traits.