Fix #8 - Reduce number of false positives inside of houses
Turns out that some doors are marked as locked inside of houses, but
there apparently is some hardcoded behavior which checks wether a player
is inside of the house and overrides the "locked" status.

The mod now uses ```door:isExteriorDoor(character)``` to check wether
a door is on the outside or the inside of a house. I tested it a bit and
it seems to work fine so far.